+++ src/main/java/com/commercecore/domain/repository/ProductRepository.java
@@ -6,13 +6,13 @@ import java.util.Optional;
 
 public interface ProductRepository {
 
-    boolean exists(Integer articleNumber);
+    boolean exists(Long articleNumber);
 
     boolean exists(Product product);
 
     void saveOrUpdateProduct(Product product);
 
-    Optional<Product> findByArtNumber(Integer articleNumber);
+    Optional<Product> findByArtNumber(Long articleNumber);
 
     void deleteProduct(Product product);
